Refactored: cleanup IDE warnings and namespace warnings
