+++ src/ContractGenerator/ContractGenerator.cs
@@ -1,7 +1,4 @@
-using System.Text;
-using Google.Protobuf;
 using Google.Protobuf.Compiler;
-using Google.Protobuf.Reflection;
 
 namespace ContractGenerator;
